Split types.go into more pieces.
